+++ CHANGELOG.md
@@ -25,6 +25,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [9.4.0] 2023-04-11
+
+### Added
+
+- Added `ImageLoader` component
+
+### Changes
+
+- Added preview options for `DragFileUploadInput`
+
 ## [9.3.2] 2023-04-06
 
 ### Changes